added an adapter
